+++ src/osrs_toolkit/__init__.py
@@ -1,3 +1,3 @@
 """OSRS Toolkit package."""
 
-__version__ = "1.0.2"
+__version__ = "1.1.0"